+++ examples/rev-list.c
@@ -24,7 +24,7 @@ int main (int argc, char **argv)
 	git_oid oid;
 	char buf[GIT_OID_HEXSZ+1];
 
-	git_threads_init();
+	git_libgit2_init();
 
 	check_lg2(git_repository_open_ext(&repo, ".", 0, NULL), "opening repository", NULL);
 	check_lg2(git_revwalk_new(&walk, repo), "allocating revwalk", NULL);
@@ -36,7 +36,7 @@ int main (int argc, char **argv)
 		printf("%s\n", buf);
 	}
 
-	git_threads_shutdown();
+	git_libgit2_shutdown();
 	return 0;
 }